+++ core/api.py
@@ -1,5 +1,6 @@
 # Api
 
+import inspect
 from parse import parse
 from webob import Request, Response
 
@@ -44,7 +45,14 @@ class API:
         handler, kwargs = self.find_handler(request_path=request.path)
 
         if handler is not None:
-            handler(request, response, **kwargs) # kwargs injects the named parameters from the parse library
+            if inspect.isclass(handler):
+                handler_function = getattr(handler(), request.method.lower(), None)
+                if handler_function is None:
+                    raise AttributeError("Method not allowed", request.method)
+                
+                handler_function(request, response, **kwargs)
+            else:
+                handler(request, response, **kwargs) # kwargs injects the named parameters from the parse library
         else:
             self.default_response(response)