+++ core/app.py
@@ -27,3 +27,11 @@ def greeting(request, response, name):
 def sum(request, response, num_1, num_2):
     total = int(num_1) + int(num_2)
     response.text = f"{num_1} + {num_2} = {total}"
+
+@app.route("/book")
+class BooksResource:
+    def get(self, req, resp):
+        resp.text = "Books Page"
+    
+    def post(self, req,resp):
+        resp.text = "Endpoint to create a book"